Split layout command handlers by concern
