+++ internal/server/commands_layout.go
@@ -698,190 +698,6 @@ func parseCopyModeArgs(args []string) (copyModeOptions, error) {
 	return opts, nil
 }
 
-func cmdNewWindow(ctx *CommandContext) {
-	var name string
-	for i := 0; i < len(ctx.Args)-1; i += 2 {
-		if ctx.Args[i] == "--name" {
-			name = ctx.Args[i+1]
-		}
-	}
-	activePid, _, _, err := ctx.activeWindowSnapshot()
-	if err != nil {
-		ctx.replyErr(err.Error())
-		return
-	}
-	meta := mux.PaneMeta{Dir: mux.PaneCwd(activePid)}
-	ctx.replyCommandMutation(ctx.Sess.enqueueCommandMutation(func(sess *Session) commandMutationResult {
-		w := sess.activeWindow()
-		if w == nil {
-			return commandMutationResult{err: fmt.Errorf("no session")}
-		}
-		pane, err := sess.createPaneWithMeta(ctx.Srv, meta, w.Width, mux.PaneContentHeight(w.Height))
-		if err != nil {
-			return commandMutationResult{err: err}
-		}
-
-		winID := sess.windowCounter.Add(1)
-		newWin := mux.NewWindow(pane, w.Width, w.Height)
-		newWin.ID = winID
-		newWin.LeadPaneID = pane.ID
-		if name != "" {
-			newWin.Name = name
-		} else {
-			newWin.Name = fmt.Sprintf(WindowNameFormat, winID)
-		}
-		sess.Windows = append(sess.Windows, newWin)
-		sess.ActiveWindowID = winID
-
-		return commandMutationResult{
-			output:          fmt.Sprintf("Created %s\n", newWin.Name),
-			broadcastLayout: true,
-			startPanes:      []*mux.Pane{pane},
-		}
-	}))
-}
-
-func cmdSelectWindow(ctx *CommandContext) {
-	if len(ctx.Args) < 1 {
-		ctx.replyErr("usage: select-window <index|name>")
-		return
-	}
-	ref := ctx.Args[0]
-
-	ctx.replyCommandMutation(ctx.Sess.enqueueCommandMutation(func(sess *Session) commandMutationResult {
-		w := sess.resolveWindow(ref)
-		if w == nil {
-			return commandMutationResult{err: fmt.Errorf("window %q not found", ref)}
-		}
-		sess.activateWindow(w)
-		return commandMutationResult{
-			output:          "Switched window\n",
-			broadcastLayout: true,
-			paneRenders:     activePaneRender(w),
-		}
-	}))
-}
-
-func cmdNextWindow(ctx *CommandContext) {
-	ctx.replyCommandMutation(ctx.Sess.enqueueCommandMutation(func(sess *Session) commandMutationResult {
-		sess.nextWindow()
-		return commandMutationResult{
-			output:          "Next window\n",
-			broadcastLayout: true,
-			paneRenders:     activePaneRender(sess.activeWindow()),
-		}
-	}))
-}
-
-func cmdPrevWindow(ctx *CommandContext) {
-	ctx.replyCommandMutation(ctx.Sess.enqueueCommandMutation(func(sess *Session) commandMutationResult {
-		sess.prevWindow()
-		return commandMutationResult{
-			output:          "Previous window\n",
-			broadcastLayout: true,
-			paneRenders:     activePaneRender(sess.activeWindow()),
-		}
-	}))
-}
-
-func cmdRenameWindow(ctx *CommandContext) {
-	if len(ctx.Args) < 1 {
-		ctx.replyErr("usage: rename-window <name>")
-		return
-	}
-	name := ctx.Args[0]
-	ctx.replyCommandMutation(ctx.Sess.enqueueCommandMutation(func(sess *Session) commandMutationResult {
-		w := sess.activeWindow()
-		if w == nil {
-			return commandMutationResult{err: fmt.Errorf("no window")}
-		}
-		w.Name = name
-		return commandMutationResult{
-			output:          fmt.Sprintf("Renamed window to %s\n", name),
-			broadcastLayout: true,
-		}
-	}))
-}
-
-func cmdResizeBorder(ctx *CommandContext) {
-	if len(ctx.Args) < 3 {
-		ctx.replyErr("usage: resize-border <x> <y> <delta>")
-		return
-	}
-	x, err1 := strconv.Atoi(ctx.Args[0])
-	y, err2 := strconv.Atoi(ctx.Args[1])
-	delta, err3 := strconv.Atoi(ctx.Args[2])
-	if err1 != nil || err2 != nil || err3 != nil {
-		ctx.replyErr("resize-border: invalid arguments")
-		return
-	}
-	ctx.replyCommandMutation(ctx.Sess.enqueueCommandMutation(func(sess *Session) commandMutationResult {
-		if w := sess.activeWindow(); w != nil {
-			w.ResizeBorder(x, y, delta)
-		}
-		return commandMutationResult{broadcastLayout: true}
-	}))
-}
-
-func cmdResizeActive(ctx *CommandContext) {
-	if len(ctx.Args) < 2 {
-		ctx.replyErr("usage: resize-active <direction> <delta>")
-		return
-	}
-	direction := ctx.Args[0]
-	delta, err := strconv.Atoi(ctx.Args[1])
-	if err != nil {
-		ctx.replyErr("resize-active: invalid delta")
-		return
-	}
-	ctx.replyCommandMutation(ctx.Sess.enqueueCommandMutation(func(sess *Session) commandMutationResult {
-		if w := sess.activeWindow(); w != nil {
-			if w.ActivePane != nil && w.IsLeadPane(w.ActivePane.ID) {
-				return commandMutationResult{err: fmt.Errorf("cannot operate on lead pane")}
-			}
-			w.ResizeActive(direction, delta)
-		}
-		return commandMutationResult{broadcastLayout: true}
-	}))
-}
-
-func cmdResizePane(ctx *CommandContext) {
-	if len(ctx.Args) < 2 {
-		ctx.replyErr("usage: resize-pane <pane> <direction> [delta]")
-		return
-	}
-	direction := ctx.Args[1]
-	switch direction {
-	case "left", "right", "up", "down":
-	default:
-		ctx.replyErr(fmt.Sprintf("resize-pane: invalid direction %q (use left/right/up/down)", direction))
-		return
-	}
-	delta := 1
-	if len(ctx.Args) >= 3 {
-		var err error
-		delta, err = strconv.Atoi(ctx.Args[2])
-		if err != nil || delta <= 0 {
-			ctx.replyErr("resize-pane: invalid delta")
-			return
-		}
-	}
-	ctx.replyCommandMutation(ctx.Sess.enqueueCommandMutation(func(sess *Session) commandMutationResult {
-		p, w, err := sess.resolvePaneWindowForActor(ctx.ActorPaneID, "resize-pane", ctx.Args[:1])
-		if err != nil {
-			return commandMutationResult{err: err}
-		}
-		if w.IsLeadPane(p.ID) {
-			return commandMutationResult{err: fmt.Errorf("cannot operate on lead pane")}
-		}
-		w.ResizePane(p.ID, direction, delta)
-		return commandMutationResult{
-			output:          fmt.Sprintf("Resized %s %s by %d\n", p.Meta.Name, direction, delta),
-			broadcastLayout: true,
-		}
-	}))
-}
-
 func parseEqualizeCommandArgs(args []string) (widths, heights bool, err error) {
 	mode := ""
 	for _, arg := range args {
